+++ auditlog/registry.py
@@ -371,6 +371,9 @@ class AuditlogModelRegistry:
                     model=model, m2m_fields=m2m_fields, exclude_fields=exclude_fields
                 )
 
+        if not isinstance(settings.AUDITLOG_STORE_JSON_CHANGES, bool):
+            raise TypeError("Setting 'AUDITLOG_STORE_JSON_CHANGES' must be a boolean")
+
         self._register_models(settings.AUDITLOG_INCLUDE_TRACKING_MODELS)